backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.9.0
chromadb==0.5.20
sentence-transformers==3.3.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
